impr(Devise): change configuration as reviewed

+++ db/migrate/20181028084301_add_devise_to_users.rb
@@ -2,16 +2,19 @@
 
 class AddDeviseToUsers < ActiveRecord::Migration
   def self.up
-    drop_table :users
+    if table_exists? :users
+      drop_table :users
+    end
 
     create_table :users do |t|
       ## Database authenticatable
-      t.string :email,              null: false, default: ""
+      # t.string :email,              null: false, default: ""
       t.string :encrypted_password, null: false, default: ""
+      t.string :username,           null: false
 
       ## Recoverable
-      t.string   :reset_password_token
-      t.datetime :reset_password_sent_at
+      # t.string   :reset_password_token
+      # t.datetime :reset_password_sent_at
 
       ## Rememberable
       t.datetime :remember_created_at
@@ -39,10 +42,11 @@ class AddDeviseToUsers < ActiveRecord::Migration
       t.timestamps null: false
     end
 
-    add_index :users, :email,                unique: true
-    add_index :users, :reset_password_token, unique: true
+    # add_index :users, :email,                unique: true
+    # add_index :users, :reset_password_token, unique: true
     # add_index :users, :confirmation_token,   unique: true
     # add_index :users, :unlock_token,         unique: true
+    add_index :users, :username,             unique: true
   end
 
   def self.down
